+++ src/lid_driven_cavity/solver_correction.rs
@@ -5,12 +5,12 @@ use super::{LidDrivenCavity, Links};
 impl LidDrivenCavity {
     pub fn solver_correction(
         &self,
-        x: &mut [f32],
-        a_0: &[f32],
+        x: &mut [f64],
+        a_0: &[f64],
         links: &[Links],
-        sources: &[f32],
+        sources: &[f64],
         iter: usize,
-        dumping: f32,
+        dumping: f64,
     ) {
         let n = self.nx;
         let mut diagonal = vec![0.0; self.nx];
@@ -257,14 +257,14 @@ impl LidDrivenCavity {
 }
 
 #[inline(always)]
-fn replace_row(row: usize, x: &mut [f32], solx: &[f32], nx: usize) {
+fn replace_row(row: usize, x: &mut [f64], solx: &[f64], nx: usize) {
     for i in 0..nx {
         x[ix(i, row, nx)] += solx[i];
     }
 }
 
 #[inline(always)]
-fn replace_column(column: usize, x: &mut [f32], soly: &[f32], nx: usize, ny: usize) {
+fn replace_column(column: usize, x: &mut [f64], soly: &[f64], nx: usize, ny: usize) {
     for j in 0..ny {
         x[ix(column, j, nx)] += soly[j];
     }
